new interpolation function for loops
A new function Alm_interp_iter() for interpolating is available. It allows to pre-load the grids for l=1,2,3 and use it to perform interpolation. This instead of loading and decompressing from disk each time the grids.
The improvment factor is of ~4 compared to the previous interpolation scheme.

Also, I added a speed test in the tests directory.