Add Alm_interp_iter_preinitialised function
The Alm_interp_iter_preinitialised() is a mean to interpol in a loop. It requires a preinitialisation of a fix-set of interpolating GSL functions (set up to lmax=3)
It is much more suitable than Alm_interp_iter which was still loading the data each time.
And it is much faster than Alm itself. The speed does not depend too much on the grid size. See the tests directory for speed tests