GridMaker makes the grid slightly larger on delta to avoid edges issues on TAMCMC
When testing on the TAMCMC application, some values very close to delta=pi/4 were giving the error: gsl: interp2d.c:147: ERROR: interpolation y value out of range
This is an edge case issue due to the fact that as the user provides the resolution and that the range is fixed, a small rounding is required to get a fixed number of grid points. There was a small error there. So now the grid is made such that it stops on delta_max = Pi/4 + resol*2